Move Job Status Functions to Sdk
