+++ src/Sdk/Pipelines/ExecutionContext.cs
@@ -1,0 +1,10 @@
+using System.Threading;
+
+namespace Sdk.Pipelines
+{
+    class ExecutionContext
+    {
+        public CancellationToken Cancelled { get; set; }
+        public JobItemFacade JobContext { get; set; }
+    }
+}
